// src/app/api/shared/emailing/emailing-send-request.interface.ts
export interface EmailingSendRequestInterface {
  sender_id: string;
  promotions: string[];
  properties: string[];
  language: string;
  subject: string;
  message: string;
  email_ready_made_content: string;
  copy_to_sender: string;
  copy_to_related_brokers: string;
  lead_id?: string[];
  close_the_lead?: string;
  in_reply_to?: string;
  hide_price: string;
  reminder: string;
  contact_id: string[];
  agency_id: string[];
  email: string[];
  firstname: string[];
  lastname: string[];
  cc_contact_id: string[];
  cc_agency_id: string[];
  cc_email: string[];
  cc_firstname: string[];
  cc_lastname: string[];
  attachments: string[];
  brochure_type: string;
  brochure_template: string;
  email_template: string;
  brochure_broker_id: string;
  reminder_contact_id: string[];
  reminder_agency_id: string[];
  reminder_email: string[];
  reminder_firstname: string[];
  reminder_lastname: string[];
  reminder_date: string;
  reminder_time: string;
}